add player pos randomize, idx <-> real convert

+++ src/Render.cpp
@@ -8,7 +8,7 @@ void Render::playerRender(Player& p, Color color){
 void Render::backgroundRender(Color color){
     ClearBackground(color);
 }
-void Render::mapRender(map_t& map, const int block_size, const uint8_t map_start_pos_x, const uint8_t map_start_pos_y){
+void Render::mapRender(map_t& map, const float block_size, const uint8_t map_start_pos_x, const uint8_t map_start_pos_y){
     for(int i=0; i<map.size(); ++i){
         for(int j=0; j<map[i].size(); ++j){
             if(map[i][j] == 0){
